Renamed: component PickerSortingType.vue
