+++ apps/frontend/src/components/page-products/PickerSortingType.vue
